fixed search issue reported by @maestrosoft (#2)
by reverting to an earlier version of MMC_insertAndFindBestMatch()
that do not have the reported issue.

It seems that the main difference is the  update logic
the new (buggy) version was using lazyUpdate()
while the older one makes direct invocation of MMC_insert_once().

I suspect it might cause issues when skipping over sections of input,
but when dutifully inserting all positions one by one,
the old version seems to work fine.

So revert to that.
One will need a better test system before investigating again such issues and optimizations.